refactor(diag): Pull the parse pass out of workspace
